+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: tmarx <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/28 14:51:47 by tmarx             #+#    #+#             */
-/*   Updated: 2019/10/29 15:28:02 by tmarx            ###   ########.fr       */
+/*   Updated: 2019/11/04 09:40:51 by tmarx            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int				min(int a, int b)
 	return (b);
 }
 
-unsigned int	abs_(int a)
+unsigned int	abs_(long int a)
 {
 	if (a < 0)
 		return (-a);
